+++ src/ui/input/index.ts
@@ -2,5 +2,6 @@ export * from "./Checkbox";
 export * from "./Field";
 export * from "./Input";
 export * from "./IntegerInput";
+export * from "./RadioGroup";
 export * from "./Select";
 export * from "./ValueInput";